cassandra: Fix multiple invocations of GET script
Use "mkdir -p" so that the GET script does not fail if the "upstream"
directory already exists. Also use the "-nc" (no clobber) option for
wget to avoid downloading files tha are already there.

This makes it possible to run GET multiple times which is useful if you
want to tinker with Cassandra configuration, for example.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>
Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>