Fix missing tzdb.dat in openjdk8-fedora module
openjdk8-fedora's "java.so" will not work because
jre/lib/tzdb.dat is left as a broken link to
/usr/share/javazi-1.8/tzdb.dat.

This patch fetches this file from Fedora's tzdata-java
package. It actually fetches and opens the entire
tzdata-java package, but because we don't copy any of
it in module.py, only the tzdb.dat file will get
copied. Perhaps in the future we'll need to copy more.
But "java.so -version" already works just by adding
the missing tzdb.dat.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>