rpm: provide our own tool
This patch provides an amendment to the rpm package, whereas our own
tools/rpm.so is used instead of the rpm binary. This tool will better serve our
needs, and in specific, allow us to better capture the output of commands at
this point: rpm will only write to standard output, and changing that from the
command line requires a pipe/fork combination. In the future, we can also enhance
this utility in new ways that may serve us.

I have considered leaving the rpm module alone, and creating a new module for
the tool. However, the utility of an rpm binary outside the realm of our
application installation system is questionable, since rpm is not a server
software anyway.  So I am just modifying it here.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>