apps/node - enable use as module during build
Enable "make module" and "make clean" to work as expected by the build
system, so that running "make image=node" in the top-level OSv directory
will build a combined kernel+node image.

The "node" module only has the node.so base, and no demo or framework
running on top of it. See the "node-demo" module built on "node" as
a demo.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>