Added apps/node-demo module
The "node-demo" module is a simple Node.JS module. It builds on the "node" module
(which is the basic node.js binary) and adds to it the "express" framework and a
silly example web server (listening on port 3000) printing "Hello World".

Perhaps we can replace this by a more impressive demo in the future?

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>