Fix memcached package
Apparently, starting with gcc 4.9, a subtle change happened in PIE
compilation - unused symbols are no longer exported in the dynamic symbol
table. This of course makes a lot of sense if you view a PIE as an
executable, but makes less sense if you pretend it is a shared library,
like we do - to run a PIE we look for its main(), and now it will no
longer be found in the dynamic symbolo table.

The trivial workaround which I do here is to add "-rdynamic" to the
compilation.

The real soulution would be to implement what is described in issue #350.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>