Fix rogue app crash
When compiled on my system (using gcc 4.9.1 from Fedora rawhide,
and the "xterm" terminal definition from Fedora 20), the rogue app
crashes on startup.

Turns out that this is a bug in the ncurses library - which in certain
cases uses a variable "j" before being initialized. I guess until now
we were lucky enough to get a low j not causing a crash, but now I got
some high random j causing the application to crash. I sent a patch to
the ncurses developers, but for now I'm also adding this patch to this
app's build, so that the rogue demo goes back to working order.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>