redis: disable setproctitle() code
Redis uses an implementation of setproctitle() which assumes that arguments
are immediately followed by environment variables, which isn't true in
case of OSv. Since, that implementation of setproctitle() won't have any
effect on OSv anyway, the easiest solution is to disable it.

Original patch by Pawel Dziepak, modified by Nadav Har'El to undefine
a couple of macros that already existed, instead of new #ifdef's to
comment out code.

Signed-off-by: Pawel Dziepak <pdziepak@quarnos.org>
Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>