openjdk8-fedora: stop slowing down "make clean"
The openjdk8-fedora module's makefile has some variables which, when used,
need to fetch the list of packages on Fedora's site. This is extremely slow -
can take 10-30 seconds in my experience. This would be perfectly fine if
it only happened during "make", but as the Makefile was currently written,
because these variables were always evaluated (as part of rules' target!)
this slow lookup was done even when someone does "make clean".

This made OSv's top-level "make clean" very slow (hanging for a long time
on the "MODULES CLEAN" part), for no good reason.

Fixed this problem by changing openjdk8-fedora's makefile not to have
target names which depend on these slow variables. Instead just inline
the wget code. "make clean" is back to reasonable speed.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>