Remove "openjdk8" module
The "openjdk8" compiles openjdk8 from source. This was useful before
OpenJDK was officially released, and there were no binary distributions
to use. Today, it is much easier and faster to just take a pre-compiled
OpenJDK8, as we do in the openjdk8-fedora module, for example.

Compiling this openjdk8 module was always excruciatingly slow and had
the user needed to install dozens of dependencies first, but today, it
no longer even compiles for me. So instead of trying to figure out why,
it's time to just say bye-bye to this module, and save our users the
misfortune of encountering it in the future and wasting their time on
compiling Java when OSv can just as well run pre-compiled Linux code.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>