Fix apps/dropwizard-example
This app no longer built, because the name of the jar file changed.
So this patch updates the name.

While we're at it, also tell Maven not to run tests on the downloaded
packages. These tests take ages, and done on every image build (even
if the packages aren't downloaded again) which is very annoying...

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>