ImageMagick example
Add an "ImageMagick" module which is an example of using the host's
unmodified ImageMagick libraries in OSv.

This trivial example (which only runs "convert --help") seems to work
when the libraries are taken from Fedora 21, but in the past we've seen
problems in libgomp (see osv issues #589 and #590) which may make this
fail on other distributions.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>