Go application example, using gccgo
This is a trivial "hello world" application in Go, which we compile with
gccgo (the traditional go compiler still doesn't work for OSv, because it
generates a static executable we still cannot run).

Thanks to Franco Venturi for doing most of the work on this example.

The example "almost" works correctly - it "usually" produces the following
output:

    $ scripts/run.py
    OSv v0.18-18-ge867864
    eth0: 192.168.122.15
    sched_getaffinity() stubbed
    sigaltstack() stubbed
    Hello, <something-in-chinese>
    Go version: xgcc (GCC) 4.9.2 20150212 (Red Hat 4.9.2-6)
    Assertion failed: timeout == 0 (/home/nyh/osv/linux.cc: futex: 55)
    QEMU: Terminated

We still need to understand what futex feature is missing and causing the
crash at the end. Also, once in a few runs we get different types of
earlier crashes, such as:

    WARNING: unimplemented void __stack_chk_fail(void)
    Aborted

And these need to be debugged. We also need to check what Go's "split
stack" means for OSv, and whether we need to use -fno-split-stack
somehow.

NOTE: Unfortunately, gccgo's main(), included in libgobegin.a, was not
compiled with -fPIC so we can't use it in OSv. Fortunately, however,
all this main() does is call several functions in sequence - so this
example includes a tiny go-main.c, a trivial clone of the main()
included in libgobegin.a.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>