perl-hello: support for scripts/build
Add support to build perl-hello with "scripts/build image=perl-hello", and
not just Capstan. The file module.py specifies that this module depends on
the "perl" module (so it gets built, and included in the image) and the
command line, and the file usr.manifest specifies which additional files
to include in the image (in this case, just hello.pl).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>