openjdk8-fedora: fix build
The build of the openjdk8-fedora was broken, at least on two systems I
tried - when it uploaded the files, it tried to create a symbolic link
on top of an existing directory, which failed and hung the build.

The problem was that the order of lines in the manifest file (and related
module.py instructions) is not guaranteed, so the creation of a symbolic
link to a directory was reordered *after* that link was used as a directory
name (and when this was done, the non-existant directory was created).

Since it is really hopeless to make any guarantees of the ordering of
manifest operations, the simpler thing to do is just to create the desired
links inside the directory which will be uploaded recursively. So this
patch moves the link creation from "usr.manifest" and "module.py" to the
Makefile, which creates that directory to be uploaded.

With this patch, "scripts/build image=openjdk8-fedora" is back to working
on my system.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>