netperf: fix re-make after failed compilation
In our netperf Makefile, if compilation once failed (e.g., the upstream
source code could not be fetched), the "upstream" directory will remain,
and any previous attempt to build netperf will use this as a sign that
nothing need to be built. Moreover, the non-existant usr.manifest will
mean that "scripts/build image=netperf" will silently build an image
with no application on it.

The solution is to change the Makefile to use "usr.manifest" as a sign
that everything has been built - instead of the "upstream" directory.
The "usr.manifest" file is the last thing we built, so if it exists,
it means the previous build was successful.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>