ruby: update version to 2.2.4
Update the ruby version downloaded by the "ruby" app from 2.2.3 to 2.2.4.

The added benefit is that we can now compile this app on Fedora 23, which
includes this specific version of Ruby (it is important because
unfortunately the current apps/ruby build requires the same version of
Ruby to be installed on the host...).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>