Drop old version of mysql
In the past, we had problems with running recent MySQL on OSV, so we
also had the "mysql41" module with an ancient MySQL 4.1 release.
This ancient version no longer compiles on modern distributions (see
https://bugs.mysql.com/bug.php?id=71250), so it's time to drop it.

Let's rename the only remaining mysql module (currently set to download and
compile version 5.6.21) to be simply "mysql".

Signed-off-by: Nadav Har'El <nyh@scylladb.com>