Go application example, using golang
This is a trivial "hello world" application in Go, which we compile with
the traditional golang compiler (see the separate gccgo-example app for
an example with the gccgo compiler).

This example does *not* yet work at this stage, but it is useful to have
a common ground for discussion on Go support in OSv.

I used the new "c-shared" build mode of golang. I don't know if it the
right thing to use but it sounds a reasonable start (I'll try later other
modes to see if they give better results).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>