apps: trivial (and non-working) fpc example application
This is a trivial hello-world Free Pascal application.
It does't currently work because of missing syscall instruction support.
https://github.com/cloudius-systems/osv/issues/590
Hopefully it will begin working when OSv adds syscall instruction support.

Stack trace we get in fpc initialization function:

  4  0x0000000000482754 in invalid_opcode (ef=0xffff800002a2b088)
     at arch/x64/exceptions.cc:320
  5  <signal handler called>
  6  0x0000100000c07b72 in SYSTEM_$$_FPSYSCALL$INT64$INT64$INT64$$INT64 ()
  7  0x0000100000c0839a in SYSTEM_$$_FPGETRLIMIT$LONGINT$PRLIMIT$$LONGINT ()
  8  0x0000000c000ff3b0 in ?? ()
  9  0x0000100000c2a357 in SYSTEM_$$_CHECKINITIALSTKLEN$QWORD$$QWORD ()
  10 0x0000000000000000 in ?? ()

Note that to see these function names, you'll need to remove the strip
command from the script ppash.sh automatically generated by fpc, and run
it again.

This example was based on
http://wiki.lazarus.freepascal.org/Free_Pascal_on_OSv
but I wanted to not need the special linking instructions.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472143611-31621-1-git-send-email-nyh@scylladb.com>