openjdk8-fedora: Add missing libraries to the image
Some of the shared libraries included with Java, such as libsunsec.so
(for security) or libjavalcms.so (for color manipulation) have their
own dependencies, additional libraries we need to download from the Fedora
distribution and include in the image.

Before this patch,

    scripts/build image=openjdk8-fedora,tomcat
    scripts/run -V

would fail because of missing libraries. Now it still fails, but because of
a different bug.

Refs #601.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>