Add lighttpd application
Add lighttpd application, compiled from source code, and some demo
configuration and index file.

Inspired by https://github.com/tari/osv-lighttpd, but uses quite
different techniques to force libtool to use "-shared" (I didn't want
to make a copy of the long compilation line), and for finding libraries
to include in the image.

Note that this application uses libpcre.so, and currently, it takes it
from the build machine. So this application may not be able to run in the
same ELF namespace as the httpserver server application or anything which
uses module/openssl, because that too includes libpcre.so.1, but not the
one in the build machine but rather the one from external/.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170709135412.18502-1-nyh@scylladb.com>