Fix compilation of iperf module
One of the tests they do with autoconf redefines bool to int (because C
does not have "bool") but then uses it in C++, which causes compilation
errors with the standard templates. There is simply no reason for this
test - none of the C files actually use "bool"! So just remove it.

Fixes #55.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>