wrk: allow larger file descriptor value
Commit c8b1906cf59d73a7cb39926c53dc4e136611d4f5 was broken, as it
allowed use of allocated array beyond its end. File descriptor value is
used as index into array. Array is allocated to be setsize large, and
check for "fd > setsize" prevents out-of-bound usage. Removing the check
was rather dumb :/

With this patch, the check remains in place. Instead, some extra memory
is allocated, so wrk is less likely to fail.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <20170904201649.13341-3-justin.cinkelj@xlab.si>