apps/apache-derby: fix Makefile
Depending on your luck, apache-derby's make might fail because it has two
parallel rules which need ROOTFS/ to exist, but only one of them creates it.
Remember when one target has two dependencies the rules are not necessarily
executed in the order the dependencies were listed! On my machine, a
"make module" simply failed.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>