apps/memcached: switch from PIE to shared object
Unfortunately, because the way recent compiler use copy relocations
for global variables, optarg (from getopt(3)) no longer works correctly
when the executable is built as a PIE. The whole story is explained in
issue #689. So let's switch to building a shared object, so the memcache
application will work again even with modern compilers.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>