apps/rogue: fix compilation
The rogue application compiles ncurses 5.9, but incorrectly used
header files from the system. If those are too new, rogue cannot
compile. So fix the Makefile to use the header files we downloaded
from ncurses 5.9.

Also use the opportunity to go back to downloading rogue from
rogueforge.net, which is back online, instead of a mirror.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>