rogue: fix terminfo
We already know that QEMU has a bug which wrongly translates a bare linefeed
sent by the guest (intentionally!) to \n\r - this bug is known for 11 years
now, see

	https://bugs.launchpad.net/qemu/+bug/1407813
	https://bugs.launchpad.net/qemu/+bug/1715296

So if we take the host's terminfo files and want to use it in the guest, we
need to avoid using \n for moving the cursor directly down - because in qemu
it doesn't (the extra \r causes the cursor to move to the *beginning* of that
line).

We had a sed script to use something else instead of \n, but it assumed
the file has the text "^J" and doesn't work if the file had "\n" - as it
now does in Fedora 27 for example. So update the sed script to allow both.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>