rust-example: replace use of "command" in Makefile
I don't think there's a real need to use the "command -v" command in
this makefile - for example if we just ignore the results from "cargo
clean" (with a minus) we don't need to check first if it exists.

Note that "command" is only a builtin in bash and it does not exist, for
example, in dash (Ubuntu's default /bin/sh). Nevertheless, even systems
with dash may have "command" as a stand-alone program in /usr/bin.
Ironically, bash installs a /usr/bin/command which uses bash and runs its
builtin "command" command.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180604120802.15958-1-nyh@scylladb.com>