openjdk8-fedora: fix latest.sh
The latest.sh script is supposed to return just one package, and the Makefile
fails if it doesn't. But if there are multiple versions in the Fedora
repository (and currently, I don't know why, there are multiple versions
of the "nss" package) it returns more than one. So fix it to return just one.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>