python2x: also copy stuff from /usr/lib64
On Fedora 28, there is both /usr/lib/python2.7 and /usr/lib64/python2.7,
and stuff from both directories need to be copied to the image to work.
So change the script to loop over
    ${PYTHON_PREFIX_DIR}/lib*/python${PYTHON_VERSION}/
instead of just the one directory
    ${PYTHON_PREFIX_DIR}/lib/python${PYTHON_VERSION}/

All these files are copied by the current script into one directory, /lib.
This seems to be fine also on Fedora 28 - it finds the files even though
they are in the wrong directory - so I didn't change this aspect of the
script.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>