add "unknown-term" package
Add an "unknown-term" package which provides /usr/share/terminfo/u/unknown,
i.e., a terminal definition for the TERM=unknown.

Note that although ncurses defaults to TERM=unknown even if TERM is unset
(as it is by default, in OSv), unfortunately some applications such as
python don't an unset TERM well, and so TERM needs to be set to "unknown"
explicitly.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>