python2x: fix terminal setting
Before this patch, trying to run python interactively ("run.py -e /python")
results in backspaces not working (outputting spaces instead).

It turns out that for backspaces to work, we need to set TERM, and need
that terminal definition to exist. In this patch we require the "unknown-term"
package - which defines the "unknown" terminal.

Sadly, while terminfo can handle an unset TERM environment variable - as is
the default on OSv - and fall back to the name "unknown", Python cannot do
this and requires TERM=unknown to be explicitly set. So in this patch we
also add a default command line of "--env=TERM=unknown /python".
As a convenience, we add that command line as the default command line
of the python2x package.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>