Fix gccgo-example
Changes to the gccgo runtime library made it necessary to change the main()
starting a go program, and the gccgo-example app now longer works.

However, since libgobegin.a is now compiled as -fPIC (see
https://patchwork.ozlabs.org/patch/464423/), we can take main() directly
from that library, instead of needing to write it on our own, and everything
becomes simpler (and works).

Fixes https://github.com/cloudius-systems/osv/issues/693

Signed-off-by: Nadav Har'El <nyh@scylladb.com>