Add "java8" default module
This patch was suggested and prepared by Geraldo Netto <geraldonetto@gmail.com>

We have various modules (e.g., openjdk8-fedora, openjdk8-zulu-*) that
provide() the "java8" feature. Packages which need Java 8 can require()
this feature. But so far we had no *default* java8 package, so to build
a package requiring java 8 it was necessary to pass one of the java8-
providing packages on the scripts/build line.

As Java 8 is becoming more and more useful (e.g., Tomcat no longer runs
on Java 7), it's becoming more convenient to choose one default
implementation of it, so just "scripts/build image=tomcat" can work.

In this patch we choose the default java8 implementation to be
openjdk8-zulu-full. The user can of course override this choice by
specifying a different Java 8 implementation in the scripts/build
line, for example

   scripts/build image=openjdk8-fedora,tomcat

will take openjdk8-fedora, instead of the default openjdk8-zulu-full.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>