openjdk8-zulu-full: fix build when new openjdk version comes out
Our Makefile uses "wget -c" to avoid re-downloading the very large openjdk
package on every build, or allow the build to continue if interrupted in
the middle of the long download.

However, "wget -c" can only be used to retrieve the *same* file again.
We actually retrieve the latest version, which may be a completely different
file... So "wget -c" may fail or result in an unusable file.

What we need to do is to name the target of the download with the same
unique name of the version we are downloading. If a new version comes
out, "wget -c" will download it from scratch because the target file
does not yet exist. The older download will remain unused, and all of
the downloads will be removed eventually by a "make clean".

Signed-off-by: Nadav Har'El <nyh@scylladb.com>