fix openjdk8-fedora
This patch fixes a crash reported by issue #880. The crash
actually happens due to a missing library libfreeblpriv3.so
provided by nss-softokn-freebl package. The root cause
was discovered by tracing vfs* operation where libsunec
would try and fail to dlopen() that library.

It turns out that newer versions of the package java-1.8.0-openjdk-headless
(for fedora 32 for example) come with libsunec that is NOT dependant
on nss* libraries (at least not dynamically) and would work
without this patch.

This patch also fixes cpio related issue that would happen
when trying to build this app on Ubuntu (see https://github.com/cloudius-systems/osv/issues/880#issuecomment-307590067
for details).

Fixes #880

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>